[feat/#123] - review_like 테이블에서 query사용하여 데이터 불러오기
