+++ src/query/review/likeQueryKeys.ts
@@ -1,0 +1,5 @@
+export const LIKED_QUERY_KEY = ['liked_data'];
+
+export const LIKED_COUNT_QUERY_KEY = ['liked_count_data'];
+
+export const LIKE_TOGGLE_KEY = ['like_toggle'];
