+++ src/query/review/reviewQueryKeys.ts
@@ -2,10 +2,4 @@ export const AUTHOR_QUERY_KEY = ['author_data'];
 
 export const REVIEW_QUERY_KEY = ['review_data'];
 
-export const LIKED_QUERY_KEY = ['liked_data'];
-
-export const LIKED_COUNT_QUERY_KEY = ['liked_count_data'];
-
-export const LIKE_TOGGLE_KEY = 'like_toggle';
-
 export const COMMENT_QUERY_KEY = ['comment_data'];
